Fix window lag with mouse raw input
For some reason, mouse raw input seems to be interfering with the system's default responses to window elements
such as the title bar and menu. For now, the code for registering the mouse for raw input is moved into a separate
function that's currently unused until needed in the future.