Check for already running instances immediately
The goal of this commit is to avoid unnecessary work if the user doesn't want and extra instance of the program.
It does so by postponing all initilization tasks until the current instance is approved by the user to run.
The only downside here is that the dialog box that prompts for launching another instance will appear before setting
the DPI awareness of the process, which results the contents of it being slightly blurred on a high-DPI monitor.
There doesn't seem to be any workarounds, since Windows allows setting process-wide DPI awareness only once.